split fastify and find-my-way code
